Experiment/requirement.txt
requests==2.31.0
python-dotenv==1.0.1
jwcrypto==1.5.6
Flask==3.0.3